Move AuthContext interface to Auth
