+++ src/auth/Auth.ts
@@ -1,6 +1,33 @@
-import { AuthContext, User } from "../shared/types";
+import { User } from "../shared/types";
 import { createContext, useContext } from "react";
 import Cookie from "js-cookie";
+
+export interface AuthContext {
+  user: User | null;
+  csrfToken: string | null;
+  check: (force?: boolean) => Promise<void>;
+  isAuthenticated: () => boolean;
+  isEditor: () => boolean;
+  post: <T>(
+    endpoint: string,
+    data: T,
+    setCurUser?: boolean
+  ) => Promise<User | undefined>;
+  put: <T>(
+    endpoint: string,
+    data: T,
+    setCurUser?: boolean
+  ) => Promise<User | undefined>;
+  patch: <T>(
+    endpoint: string,
+    data: T,
+    setCurUser?: boolean
+  ) => Promise<User | undefined>;
+  delete: (endpoint: string) => Promise<User | undefined>;
+  setUser: (user: User | null) => void;
+  login: (username: string, password: string) => Promise<User | undefined>;
+  logout: () => Promise<void>;
+}
 
 export interface AuthState {
   user: User | null;
