+++ src/shared/types.ts
@@ -117,30 +117,3 @@ export type UserAPIResponse = APIResponse<
 export interface IssueAPIError extends APIError {
   non_field_errors?: string[];
 }
-
-export interface AuthContext {
-  user: User | null;
-  csrfToken: string | null;
-  check: (force?: boolean) => Promise<void> | undefined;
-  isAuthenticated: () => boolean;
-  isEditor: () => boolean;
-  post: <T>(
-    endpoint: string,
-    data: T,
-    setCurUser?: boolean
-  ) => Promise<User | undefined>;
-  put: <T>(
-    endpoint: string,
-    data: T,
-    setCurUser?: boolean
-  ) => Promise<User | undefined>;
-  patch: <T>(
-    endpoint: string,
-    data: T,
-    setCurUser?: boolean
-  ) => Promise<User | undefined>;
-  delete: (endpoint: string) => Promise<User | undefined>;
-  setUser: (user: User | null) => void;
-  login: (username: string, password: string) => Promise<User | undefined>;
-  logout: () => Promise<void>;
-}
